# substationDataServer/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Public Substation API for the Smart Service Intelligent Autonomous Grid",
    "description" : "Public API for registered users to access the device data, production data on a device, bay or substation base.",
    "contact" : {
      "name" : "Istvan Mudrak",
      "url" : "https://www.smartservice.energy",
      "email" : "istvan.mudrak@smartservice.energy"
    },
    "license" : {
      "name" : "Proprietary",
      "url" : "https://www.smartservice.energy/licenses/Substation"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.smartservice.energy/substation/v1"
  } ],
...
  "paths" : {
...
    "/substations/{substationId}/bays" : {
      "get" : {
        "tags" : [ "bay" ],
        "operationId" : "findBayBySubstationId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "substationId",
          "in" : "path",
          "description" : "Id of a given substation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All bay data idetified by {substationId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bays"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/substations/{substationId}/devices" : {
      "get" : {
        "tags" : [ "device" ],
        "operationId" : "findDeviceBySubstationId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "substationId",
          "in" : "path",
          "description" : "Id of a given substation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All device data idetified by {substationId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Devices"
                  }
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/bays" : {
      "get" : {
        "tags" : [ "bay" ],
        "operationId" : "findBayAll",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of all bays owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bays"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/bays/{bayId}" : {
      "get" : {
        "tags" : [ "bay" ],
        "operationId" : "findBayByBayId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "bayId",
          "in" : "path",
          "description" : "Id of a given bay",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Bay data idetified by {bayId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bay"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/bays/{bayId}/devices" : {
      "get" : {
        "tags" : [ "device" ],
        "operationId" : "findDeviceByBayId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "bayId",
          "in" : "path",
          "description" : "Id of a given bay",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All device data idetified by {bayId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Devices"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/devices" : {
      "get" : {
        "tags" : [ "device" ],
        "operationId" : "findDeviceAll",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of all device owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Devices"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{deviceId}" : {
      "get" : {
        "tags" : [ "device" ],
        "operationId" : "findDeviceByDeviceId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "deviceId",
          "in" : "path",
          "description" : "Id of a given device",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Device data idetified by {deviceId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Device"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        }
      },
      "Location" : {
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "number",
            "format" : "float"
          },
          "longitude" : {
            "type" : "number",
            "format" : "float"
          },
          "altitude" : {
            "type" : "number",
            "format" : "float"
          }
        },
        "example" : {
          "altitude" : 1.4658129,
          "latitude" : 0.8008282,
          "longitude" : 6.0274563
        }
      },
...
      "Bay" : {
        "type" : "object",
        "properties" : {
          "mrid" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "description" : {
            "type" : "string"
          }
        },
        "example" : {
          "mrid" : "mrid",
          "name" : "name",
          "description" : "description",
          "location" : {
            "altitude" : 1.4658129,
            "latitude" : 0.8008282,
            "longitude" : 6.0274563
          }
        }
      },
      "Bays" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Bay"
        }
      },
      "DeviceType" : {
        "type" : "object",
        "properties" : {
          "vendor" : {
            "type" : "string"
          },
          "hwRev" : {
            "type" : "string"
          },
          "swRev" : {
            "type" : "string"
          },
          "serNum" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "owner" : {
            "type" : "string"
          },
          "primeOps" : {
            "type" : "string"
          },
          "secondOps" : {
            "type" : "string"
          }
        },
        "description" : "Based on the IEC61850 DPL Common Data Class",
        "example" : {
          "owner" : "owner",
          "hwRev" : "hwRev",
          "swRev" : "swRev",
          "vendor" : "vendor",
          "primeOps" : "primeOps",
          "serNum" : "serNum",
          "model" : "model",
          "secondOps" : "secondOps"
        }
      },
      "Device" : {
        "type" : "object",
        "properties" : {
          "mrid" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/DeviceType"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "description" : {
            "type" : "string"
          }
        },
        "example" : {
          "mrid" : "mrid",
          "name" : "name",
          "description" : "description",
          "location" : {
            "altitude" : 1.4658129,
            "latitude" : 0.8008282,
            "longitude" : 6.0274563
          },
          "type" : {
            "owner" : "owner",
            "hwRev" : "hwRev",
            "swRev" : "swRev",
            "vendor" : "vendor",
            "primeOps" : "primeOps",
            "serNum" : "serNum",
            "model" : "model",
            "secondOps" : "secondOps"
          }
        }
      },
      "Devices" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Device"
        }
      },
...
    },
    "securitySchemes" : {
      "smartservice_auth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/authorize",
            "tokenUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/token"
          }
        }
      }
    }
  }
}